third_party/blink/web_tests/platform/generic/fast/events/file-input-hidden-in-ondrop-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS dropTarget.value is "C:\\fakepath\\abe.png"
